Removed module wrapper function from the bundles
This function is constant and can much better be added in the
bundle-loader. Previously the bundle-loader had to parse the module
string to extract the function signature, whereas the signature is the
same each time. It's much more efficient to let it just create the
signature itself, and it also works better with sourcemaps.